Split CourseStatsPage into multiple components
- Removed the old CourseStatsPage component and replaced it with a new modular structure.
- Introduced separate tab components for Overview, User Activity, Visitors, All Languages, and Course Instances to enhance maintainability and readability.
- Added constants for tab identifiers and default chart height for consistent usage across components.
- Implemented dynamic imports for tab components to optimize loading performance.
- Created a new styles file for shared styles used in the InstructionBox component.
- Updated the main CourseStatsPage to utilize the new structure and components, improving overall organization and user experience.
