+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/CourseStatsPage/constants.ts
@@ -1,0 +1,7 @@
+export const TAB_OVERVIEW = "overview"
+export const TAB_USER_ACTIVITY = "user-activity"
+export const TAB_VISITORS = "visitors"
+export const TAB_ALL_LANGUAGES = "all-languages"
+export const TAB_COURSE_INSTANCES = "course-instances"
+
+export const DEFAULT_CHART_HEIGHT = 450
